\name{makeElipseCoords}
\alias{makeElipseCoords}
\title{Make Elipse Coordinates}
\description{
Create x & y coordinates for an elipse from parameters.
\code{save}.
}
\usage{
makeElipseCoords(x0 = 0, y0 = 0, b = 1, a = 1, alpha = 0, pct.range = c(0,1), len = 50)
}
\arguments{
\item{x0}{x coordinate of center of elipse.}
\item{y0}{y coordinate of center of elipse.}
\item{b}{y axis stretch factor.}
\item{a}{x axis stretch factor.}
\item{alpha}{rotation factor.}
\item{pct.range}{percentage of the way around the elipse.}
\item{len}{number of points used to draw elipse.}
}
\seealso{
}
\examples{
}
\keyword{graphs}

