\name{plotClock}
\alias{plotClock}
\title{Plot a simple clock.}
\description{
Used to create a clock on a plot as a way to keep track of the additional 
parameter of time for use in animated movies of multiple plots.}
\usage{
plotClock(hour, minute, x0 = 0, y0 = 0, r = 1)
}
\arguments{
\item{hour}{integer specifying the position of the hour hand.}
\item{minute}{integer specifying the position of the minute hand.}
\item{x0}{number specifying the x position of the clock.}
\item{y0}{number specifying the y position of the clock.}
\item{r}{number specifying the radius of the clock.}
}
\value{
a plot of a clock
}
\keyword{graphs}

