\name{raAddArms}
\alias{raAddArms}
\title{Add Arms to a RA plot.}
\description{
.
}
\usage{
raAddArms(epsilon=.55, start=1, end=6, A.shift=0, R.shift=0, ...)
}
\arguments{
\item{epsilon}{.}
\item{start}{.}
\item{end}{.}
\item{A.shift}{.}
\item{R.shift}{.}
\item{...}{other parameters passed to lines.}
}
\seealso{
}
\examples{
}
\keyword{}

