\name{raAddAxlabs}
\alias{raAddAxlabs}
\title{Add axis labels to an RA plot.}
\description{
.
}
\usage{
raAddAxlabs(conditions=nv(c('a','b'),c('ref','obs')), normalize=T, add=TRUE, line=2)
}
\arguments{
\item{conditions}{.}
\item{normalize}{.}
\item{add}{.}
\item{line}{.}
}
\seealso{
}
\examples{
}
\keyword{}

