\name{raAddSigLines}
\alias{raAddSigLines}
\title{Add Significance Lines to an RA plot.}
\description{
.
}
\usage{
raAddSigLines(n, end=20, alpha=1e-3, mm=0, A.shift=0, plot=FALSE, ...)
}
\arguments{
\item{n}{.}
\item{end}{.}
\item{alpha}{.}
\item{mm}{.}
\item{A.shift}{.}
\item{plot}{.}
\item{...}{other parameters passed to lines.}
}
\seealso{
\code{\link{raPlot}}
}
\examples{
}
\keyword{}

