\name{raPlot}
\alias{raPlot}
\title{Generate a Ratio Average [RAy] Plot.}
\description{
.
}
\usage{
raPlot(a, b, uniques=5, normalize=FALSE,  
                   mm=0, alpha = 0.01, jitter=FALSE, jit.wgts=NULL,
                   rex=1, flat=TRUE, tail=.5, arms=.5, spine=1, border=NULL, ...)
}
\arguments{
\item{a}{.}
\item{b}{.}
\item{uniques}{.}
\item{normalize}{.}
\item{mm}{.}
\item{alpha}{.}
\item{jitter}{.}
\item{jit.wgts}{.}
\item{rex}{.}
\item{flat}{.}
\item{tail}{.}
\item{arms}{.}
\item{spine}{.}
\item{border}{.}
\item{...}{other parameters passed to plot.}
}
\seealso{
}
\examples{
}
\keyword{}

