\name{sumSortedTable}
\alias{sumSortedTable}
\author{David Schruth <dschruth@u.washington.edu>}
\title{A row-wise sum sorted cross-tabulation}
\description{
This is a table command that also performs a row-wise sum and sorts by the new column.  
}
\usage{
sumSortedTable(x, clmns)
}
\arguments{
\item{x}{list of vectors or a dataframe}
\item{clmns}{accepts 'by' object for grouping}
}
\seealso{
ledghead, table, order
}
\examples{
}
\keyword{}
