\name{truncPath}
\alias{truncPath}
\title{Truncate a filesystem path.}
\description{
Truncate by shortening the file name of the path.
}
\usage{
truncPath(path, limit=256, min.length=15)
}
\arguments{
\item{path}{system path to be checked and potentially shortened.}
\item{limit}{limit to the length of the path.}
\item{min.length}{minimum length of path.}
}
\seealso{
}
\examples{
}
\keyword{manip}


