\name{pad}
\alias{pad}
\title{Pad a vector of numerical string with zeros.}
\description{
This function helps to pad numbers on the left side with zeros so that 
they may be used to create strings used in filesystem names (for example). 
}
\usage{
pad(vect,np)
}
\arguments{
\item{vect}{a vector of strings representing numbers.}
\item{np}{number of zeros to pad to the left of the string.}
}
\seealso{
}
\examples{
}
\keyword{manip}

