\name{raAddAxLabs}
\alias{raAddAxLabs}
\title{Add axis labels to an RA plot.}
\description{
.
}
\usage{
raAddAxLabs(conditions=nv(c('a','b'),c('ref','obs')), normalize=T, add=TRUE, line=2)
}
\arguments{
\item{conditions}{.}
\item{normalize}{.}
\item{add}{.}
\item{line}{.}
}
\seealso{
\code{\link{raPlot}}
}
\examples{
}
\keyword{}

