\name{write.delim}
\alias{write.delim}
\title{Write a (tab) delimited text file.}
\description{
A simple wrapper for write.table with the same options as read.delim
}
\usage{
write.delim(df, file)
}
\arguments{
\item{df}{a dataframe.}
\item{file}{outputfile path.}
}
\seealso{
}
\examples{
}
\keyword{}

